refactor: 꾸러기 패키지 구조 수정 Resolves: #1
