+++ src/main/java/com/system/fcfs/config/AwsSQSConfig.java
@@ -1,4 +1,4 @@
-package com.system.fcfs.event.config;
+package com.system.fcfs.config;
 
 import io.awspring.cloud.sqs.config.SqsMessageListenerContainerFactory;
 import io.awspring.cloud.sqs.listener.acknowledgement.AcknowledgementOrdering;
