+++ src/main/java/com/system/fcfs/config/RedisProperty.java
@@ -1,4 +1,4 @@
-package com.system.fcfs.event.config;
+package com.system.fcfs.config;
 
 import lombok.Getter;
 import lombok.Setter;
