+++ src/main/java/com/system/fcfs/event/domain/enums/Event.java
@@ -1,4 +1,4 @@
-package com.system.fcfs.event.constant;
+package com.system.fcfs.event.domain.enums;
 
 import lombok.Getter;
