+++ src/main/java/com/system/fcfs/event/domain/enums/Role.java
@@ -1,4 +1,4 @@
-package com.system.fcfs.event.constant;
+package com.system.fcfs.event.domain.enums;
 
 public enum Role {
     USER,    // 회원
